Function yew_hooks::prelude::use_media_with_options

pub fn use_media_with_options<'hook>(
    node: NodeRef,
    src: String,
    options: UseMediaOptions
) -> impl 'hook + Hook<Output = UseMediaHandle>

This hook plays video or audio and exposes its controls with options. see use_media

Note

When used in function components and hooks, this hook is equivalent to:

pub fn use_media_with_options(
    node: NodeRef,
    src: String,
    options: UseMediaOptions,
) -> UseMediaHandle {
    /* implementation omitted */
}